requirements.txt
owlready
matplotlib
mpld3
shapely
numpy
screeninfo
tqdm
./owlready2_augmentator
